web/requirements.txt
appdirs==1.4.3
appnope==0.1.0
click==6.7
decorator==4.0.11
Django==1.10.6
django-rq==0.9.5
djangorestframework==3.6.2
elasticsearch==5.2.0
elasticsearch-dsl==5.1.0
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
redis==2.10.5
requests==2.13.0
rq==0.7.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.20
wcwidth==0.1.7
xmltodict==0.10.2
